Move version.py to src to properly include it
